Finished Autocomplete component - better handling data, positioning, style; optimized AddX components
