+++ client/src/styles/Autocomplete.scss
@@ -3,10 +3,9 @@
   width: 100%;
   border: 1px solid #ccc;
   border-radius: 4px;
-  padding: 6px;
+  padding: 0 6px;
   min-height: 40px;
   max-height: 80px;
-  cursor: text;
   background: white;
   display: flex;
   align-items: center;
@@ -35,7 +34,7 @@
 
   .chip-remove {
     cursor: pointer;
-    font-size: 14px;
+    font-size: 1rem;
     line-height: 1;
     margin-left: 0.3rem;
 
@@ -50,17 +49,17 @@
     border: none;
     outline: none;
     padding: 6px;
-    font-size: 14px;
+    font-size: 1rem;
     position: relative;
     z-index: 1;
     background: transparent;
+    cursor: text;
   }
 
   .autocomplete-menu {
-    position: absolute;
+    position: fixed;
     top: 100%;
     left: 0;
-    width: 100%;
     max-height: 250px;
     overflow-y: auto;
     background: white;
@@ -68,6 +67,11 @@
     border-radius: 4px;
     box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
     z-index: 1000;
+
+    &.top {
+      top: auto;
+      bottom: 100%;
+    }
 
     .autocomplete-item {
       padding: 8px 12px;
@@ -98,6 +102,11 @@
         color: #1976d2;
         font-weight: bold;
         text-align: center;
+      }
+
+      &.disabled {
+        cursor: not-allowed !important;
+        opacity: 0.5;
       }
     }
   }
