+++ server/src/types/user.ts
@@ -5,7 +5,8 @@ export interface IUser extends Document {
     firstName?: string,
     lastName: string,
     email: string,
-    hashedPassword: string
+    hashedPassword: string,
+    deletedAt?: Date | null
 }
 
 export interface CustomJwtPayload extends JwtPayload {
